+++ packages/ts-newskit/src/components/slices/shared/articleTileInfo.tsx
@@ -1,6 +1,10 @@
 import React from 'react';
 import { Divider } from 'newskit';
-import { ContainerInline, StyledBlock } from '../shared-styles';
+import {
+  ContainerInline,
+  StyledBlock,
+  VideoIconContainer
+} from '../shared-styles';
 import { LiveTag } from './live-tag';
 import { CustomTextBlock } from './customTextBlock';
 import { getActiveArticleFlags } from '../../../utils/getActiveArticleFlag';
@@ -88,11 +92,17 @@ export const ArticleTileInfo = ({
           )}
           {hasVideo && (
             <TileWrapper>
-              <CustomTextBlock
-                text="VIDEO"
-                stylePreset="inkContrast"
-                icon={<VideoIcon />}
-              />
+              <VideoIconContainer>
+                <CustomTextBlock
+                  text="VIDEO"
+                  stylePreset="inkContrast"
+                  icon={
+                    <VideoIcon
+                      style={{ verticalAlign: 'top', marginRight: '4px' }}
+                    />
+                  }
+                />
+              </VideoIconContainer>
             </TileWrapper>
           )}
           {label && (