+++ packages/ts-newskit/src/components/slices/shared/customTextBlock.tsx
@@ -1,6 +1,5 @@
 import React, { ReactNode } from 'react';
-import { TextBlock } from 'newskit';
-import { Wrapper } from '../shared-styles';
+import { InlineTextBlock } from '../shared-styles';
 
 export const CustomTextBlock = ({
   text,
@@ -12,15 +11,13 @@ export const CustomTextBlock = ({
   icon?: ReactNode;
 }) => {
   return (
-    <TextBlock
+    <InlineTextBlock
       typographyPreset="utilityLabel005"
       stylePreset={stylePreset ? stylePreset : 'inkBrand010'}
       as="span"
     >
-      <Wrapper>
-        {icon}
-        {text}
-      </Wrapper>
-    </TextBlock>
+      {icon}
+      {text}
+    </InlineTextBlock>
   );
 };