+++ packages/article-in-depth/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.78.16](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.78.15...@times-components/article-in-depth@3.78.16) (2023-11-27)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.78.15](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.78.14...@times-components/article-in-depth@3.78.15) (2023-11-27)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
 ## [3.78.14](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.78.13...@times-components/article-in-depth@3.78.14) (2023-11-27)
 
 **Note:** Version bump only for package @times-components/article-in-depth